refactor: lift state up from address input container
