+++ apps/escrow/frontend/src/systems/Core/components/CreateEscrow.tsx
@@ -7,8 +7,28 @@ import { AddressInputContainer } from "./AddressInputContainer"
 import { AssetInputContainer } from "./AssetInputContainer";
 
 export const CreateEscrow = () => {
+    const [users, setUsers] = useState(["", ""]);
+
+    const handleUsersChange = (event: ChangeEvent<HTMLInputElement>, userIdx: number) => {
+        const newUsers = [...users];
+        newUsers[userIdx] = event.target.value;
+        setUsers(newUsers);
+    }
+
+    const handleAddUser = (event: any) => {
+        setUsers([...users, ""]);
+    }
+
+    // TODO (FIX) the way we remove and display users this really just pops the last user from the list
+    const handleRemoveUser = (userId: number) => {
+        setUsers(users.filter((user, i) => {
+            return i !== userId;
+        }));
+    }
+
     const handleSubmit = (event: SyntheticEvent) => {
         event.preventDefault();
+        console.log(users);
     }
 
 
@@ -17,7 +37,12 @@ export const CreateEscrow = () => {
             <Card css={{ margin: "10px", bg: "$gray7", marginTop: "50px" }}>
                 <form onSubmit={handleSubmit}>
                     <Stack css={{ width: "475px", margin: "10px", alignItems: "center" }}>
-                        <AddressInputContainer />
+                        <AddressInputContainer
+                            onAddUser={handleAddUser}
+                            onRemoveUser={handleRemoveUser}
+                            onUserInfoChange={handleUsersChange}
+                            users={users}
+                        />
                         <AssetInputContainer />
                         <Button type="submit" leftIcon="PlusIcon" css={{ font: "$sans", alignSelf: "stretch" }}>Create Escrow</Button>
                     </Stack>
